refactor cmd
