+++ cmd/status.go
@@ -33,7 +33,6 @@ func runStatus(cmd *cobra.Command, args []string) error {
 
 	ctx := setup.loggerContext(cmd.Context(), false)
 
-	// Load the DAG
 	dag, err := digraph.Load(ctx, args[0], digraph.WithBaseConfig(setup.cfg.Paths.BaseConfig))
 	if err != nil {
 		logger.Error(ctx, "Failed to load DAG", "path", args[0], "err", err)
@@ -52,7 +51,6 @@ func runStatus(cmd *cobra.Command, args []string) error {
 		return fmt.Errorf("failed to retrieve current status: %w", err)
 	}
 
-	// Log the status information
 	logger.Info(ctx, "Current status", "pid", status.PID, "status", status.Status)
 
 	return nil
