+++ internal/testdata/cmd/config_startall.yaml
@@ -1,3 +1,3 @@
 port: 54322
 headless: true
-dags: /tmp/.dagu_test/config/dags
+dags: /tmp/dagu_test
